Bug fix: Fixed scratchpads being moved while in fullscreen mode
Previously, the scratchpad will be animated and moved to the new
position, even if the scratchpad is in fullscreen mode. Now we won't
animate or move it if it's in fullscreen mode, but will update the
new attach position internally. As the scratchpad leaves the fullscreen
mode it will attaches to the new position.